sql/row: move fkDeleteHelper/fkUpdateHelper to separate files
Additionally, there was some unneeded cruft in the structs, which I
removed.

Release note: None
