+++ pkg/sql/row/fk.go
@@ -17,8 +17,6 @@ package row
 import (
 	"context"
 
-	"github.com/cockroachdb/cockroach/pkg/internal/client"
-	"github.com/cockroachdb/cockroach/pkg/roachpb"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgerror"
 	"github.com/cockroachdb/cockroach/pkg/sql/privilege"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
@@ -330,146 +328,4 @@ outer:
 		}
 	}
 	return nil
-}
-
-type fkDeleteHelper struct {
-	fks         map[sqlbase.IndexID][]baseFKHelper
-	otherTables TableLookupsByID
-	alloc       *sqlbase.DatumAlloc
-
-	checker *fkBatchChecker
-}
-
-func makeFKDeleteHelper(
-	txn *client.Txn,
-	table *sqlbase.ImmutableTableDescriptor,
-	otherTables TableLookupsByID,
-	colMap map[sqlbase.ColumnID]int,
-	alloc *sqlbase.DatumAlloc,
-) (fkDeleteHelper, error) {
-	h := fkDeleteHelper{
-		otherTables: otherTables,
-		alloc:       alloc,
-		checker: &fkBatchChecker{
-			txn: txn,
-		},
-	}
-	for _, idx := range table.AllNonDropIndexes() {
-		for _, ref := range idx.ReferencedBy {
-			if otherTables[ref.Table].IsAdding {
-				// We can assume that a table being added but not yet public is empty,
-				// and thus does not need to be checked for FK violations.
-				continue
-			}
-			fk, err := makeBaseFKHelper(txn, otherTables, idx, ref, colMap, alloc, CheckDeletes)
-			if err == errSkipUnusedFK {
-				continue
-			}
-			if err != nil {
-				return fkDeleteHelper{}, err
-			}
-			if h.fks == nil {
-				h.fks = make(map[sqlbase.IndexID][]baseFKHelper)
-			}
-			h.fks[idx.ID] = append(h.fks[idx.ID], fk)
-		}
-	}
-	return h, nil
-}
-
-func (h fkDeleteHelper) addAllIdxChecks(ctx context.Context, row tree.Datums, traceKV bool) error {
-	if len(h.fks) == 0 {
-		return nil
-	}
-	for idx := range h.fks {
-		if err := checkIdx(ctx, h.checker, h.fks, idx, row, traceKV); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// CollectSpans implements the FkSpanCollector interface.
-func (h fkDeleteHelper) CollectSpans() roachpb.Spans {
-	return collectSpansWithFKMap(h.fks)
-}
-
-// CollectSpansForValues implements the FkSpanCollector interface.
-func (h fkDeleteHelper) CollectSpansForValues(values tree.Datums) (roachpb.Spans, error) {
-	return collectSpansForValuesWithFKMap(h.fks, values)
-}
-
-type fkUpdateHelper struct {
-	inbound  fkDeleteHelper // Check old values are not referenced.
-	outbound fkInsertHelper // Check rows referenced by new values still exist.
-
-	indexIDsToCheck map[sqlbase.IndexID]struct{} // List of Index IDs to check
-
-	checker *fkBatchChecker
-}
-
-func makeFKUpdateHelper(
-	txn *client.Txn,
-	table *sqlbase.ImmutableTableDescriptor,
-	otherTables TableLookupsByID,
-	colMap map[sqlbase.ColumnID]int,
-	alloc *sqlbase.DatumAlloc,
-) (fkUpdateHelper, error) {
-	ret := fkUpdateHelper{
-		indexIDsToCheck: make(map[sqlbase.IndexID]struct{}),
-	}
-	var err error
-	if ret.inbound, err = makeFKDeleteHelper(txn, table, otherTables, colMap, alloc); err != nil {
-		return ret, err
-	}
-	ret.outbound, err = makeFKInsertHelper(txn, table, otherTables, colMap, alloc)
-	ret.outbound.checker = ret.inbound.checker
-	ret.checker = ret.inbound.checker
-	return ret, err
-}
-
-func (fks fkUpdateHelper) addCheckForIndex(
-	indexID sqlbase.IndexID, descriptorType sqlbase.IndexDescriptor_Type,
-) {
-	if descriptorType == sqlbase.IndexDescriptor_FORWARD {
-		fks.indexIDsToCheck[indexID] = struct{}{}
-	}
-}
-
-func (fks fkUpdateHelper) addIndexChecks(
-	ctx context.Context, oldValues, newValues tree.Datums, traceKV bool,
-) error {
-	for indexID := range fks.indexIDsToCheck {
-		if err := checkIdx(ctx, fks.checker, fks.inbound.fks, indexID, oldValues, traceKV); err != nil {
-			return err
-		}
-		if err := checkIdx(ctx, fks.checker, fks.outbound.fks, indexID, newValues, traceKV); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (fks fkUpdateHelper) hasFKs() bool {
-	return len(fks.inbound.fks) > 0 || len(fks.outbound.fks) > 0
-}
-
-// CollectSpans implements the FkSpanCollector interface.
-func (fks fkUpdateHelper) CollectSpans() roachpb.Spans {
-	inboundReads := fks.inbound.CollectSpans()
-	outboundReads := fks.outbound.CollectSpans()
-	return append(inboundReads, outboundReads...)
-}
-
-// CollectSpansForValues implements the FkSpanCollector interface.
-func (fks fkUpdateHelper) CollectSpansForValues(values tree.Datums) (roachpb.Spans, error) {
-	inboundReads, err := fks.inbound.CollectSpansForValues(values)
-	if err != nil {
-		return nil, err
-	}
-	outboundReads, err := fks.outbound.CollectSpansForValues(values)
-	if err != nil {
-		return nil, err
-	}
-	return append(inboundReads, outboundReads...), nil
 }
